requirements.txt
aiohttp==3.8.1
aiosqlite==0.17.0
opencv-python-headless==4.5.3.56
numpy==1.21.2
paddleocr==2.4.0
psutil==5.8.0
sentry-sdk==1.5.8
schedule==1.1.0
Pillow==8.3.2
matplotlib==3.4.3
pandas==1.3.3
pyyaml==5.4.1
flask==2.0.1
torch==1.9.0
torchvision==0.10.0
seaborn==0.11.2
pyenchant==3.2.1
watchdog==2.1.6
transformers==4.11.3
networkx==2.6.3
pyDatalog==0.17.1
dash==2.0.0
plotly==5.3.1
librosa==0.8.1
spacy==3.1.3
nltk==3.6.3
scikit-learn==0.24.2
tensorflow==2.6.0
keras==2.6.0
gym==0.21.0
stable-baselines3==1.2.0
optuna==2.10.0
ray[rllib]==1.7.0
pytest==6.2.5
pytest-asyncio==0.15.1
black==21.9b0
flake8==3.9.2
mypy==0.910